chore: move route page components to dedicated folder
