+++ app/component/routepage/RouteAgencyInfo.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { createFragmentContainer, graphql } from 'react-relay';
 import get from 'lodash/get';
-import { routeShape, configShape } from '../util/shapes';
-import AgencyInfo from './AgencyInfo';
+import { routeShape, configShape } from '../../util/shapes';
+import AgencyInfo from '../AgencyInfo';
 
 function RouteAgencyInfo({ route }, { config }) {
   const agencyName = get(route, 'agency.name');
