+++ app/component/routepage/RouteScheduleContainer.js
@@ -11,21 +11,21 @@ import cx from 'classnames';
 import { dayRangePattern } from '@digitransit-util/digitransit-util';
 import { getTranslatedDayString } from '@digitransit-util/digitransit-util-route-pattern-option-text';
 import isEqual from 'lodash/isEqual';
-import { routeShape, patternShape, configShape } from '../util/shapes';
+import { routeShape, patternShape, configShape } from '../../util/shapes';
 import RouteScheduleHeader from './RouteScheduleHeader';
 import RouteScheduleTripRow from './RouteScheduleTripRow';
-import SecondaryButton from './SecondaryButton';
-import Loading from './Loading';
-import { DATE_FORMAT, RealtimeStateType } from '../constants';
-import { addAnalyticsEvent } from '../util/analyticsUtils';
-import withBreakpoint from '../util/withBreakpoint';
-import hashCode from '../util/hashUtil';
-import { getFormattedTimeDate } from '../util/timeUtils';
+import SecondaryButton from '../SecondaryButton';
+import Loading from '../Loading';
+import { DATE_FORMAT, RealtimeStateType } from '../../constants';
+import { addAnalyticsEvent } from '../../util/analyticsUtils';
+import withBreakpoint from '../../util/withBreakpoint';
+import hashCode from '../../util/hashUtil';
+import { getFormattedTimeDate } from '../../util/timeUtils';
 import RouteScheduleDropdown from './RouteScheduleDropdown';
 import RoutePageControlPanel from './RoutePageControlPanel';
-import { PREFIX_ROUTES, PREFIX_TIMETABLE } from '../util/path';
-import { isBrowser } from '../util/browser';
-import ScrollableWrapper from './ScrollableWrapper';
+import { PREFIX_ROUTES, PREFIX_TIMETABLE } from '../../util/path';
+import { isBrowser } from '../../util/browser';
+import ScrollableWrapper from '../ScrollableWrapper';
 import getTestData from './RouteScheduleDebugData';
 
 const DATE_FORMAT2 = 'D.M.YYYY';
