+++ app/component/routepage/TripStopListContainer.js
@@ -7,10 +7,10 @@ import connectToStores from 'fluxible-addons-react/connectToStores';
 import groupBy from 'lodash/groupBy';
 import values from 'lodash/values';
 import moment from 'moment';
-import { configShape, tripShape, vehicleShape } from '../util/shapes';
-import { getRouteMode } from '../util/modeUtils';
+import { configShape, tripShape, vehicleShape } from '../../util/shapes';
+import { getRouteMode } from '../../util/modeUtils';
 import TripRouteStop from './TripRouteStop';
-import withBreakpoint from '../util/withBreakpoint';
+import withBreakpoint from '../../util/withBreakpoint';
 
 class TripStopListContainer extends React.PureComponent {
   static propTypes = {
