+++ sass/_main.scss
@@ -22,7 +22,7 @@ $body-font-weight: $font-weight-medium;
 @import '../app/component/front-page';
 @import '../app/component/navigation';
 @import '../app/component/bike-park-rental-station';
-@import '../app/component/route';
+@import '../app/component/routepage/route';
 @import '../app/component/checkbox';
 @import '../app/component/toggle';
 @import '../app/component/stops-near-you';
@@ -40,7 +40,7 @@ $body-font-weight: $font-weight-medium;
 @import '~zurb-foundation-5/scss/foundation/components/forms';
 @import '../app/component/zone-icon';
 @import '../app/component/date-select';
-@import '../app/component/route-schedule-dropdown';
+@import '../app/component/routepage/route-schedule-dropdown';
 @import '../app/component/embedded-search';
 @import '../app/component/embedded-search-generator';
 @import '../app/component/from-map-modal';
